% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countNA.R
\name{count_na}
\alias{count_na}
\title{Count of missing values in one column}
\usage{
count_na(data, col_name)
}
\arguments{
\item{data}{data set to get the specific column from.}

\item{col_name}{name of the column (string input) to count missing values.}
}
\value{
A 1 x 1 tibble with variable \code{col_name} and its value \verb{integer count of missing values}.
}
\description{
The goal of \code{countNA} is to count missing values (NA) in the specified column \code{col_name} of the given data set \code{data}.
}
\examples{
# penguins is a data set from palmerpenguins that has columns "sex" and "year".
library(palmerpenguins)
count_na(penguins, "sex")
count_na(penguins, "year")
}
